Fix Invalid DateTime (#425)
* Fix Invalid DateTime

The default numbering system "latn" used to parse the time in locale string returns a Narrow No-Break Space (NNBSP) instead of standard space which cause the parse to fail. The proposed fix replace the NNBSP for a standard space.

* Replacing NNBSP with UTF Code

Replacing NNBSP with UTF Code

* Updating versions

* Numbering System parameters fix + spec

The numbering system parameter is required every time a datetime is parsed. Previously, the parameters used default language which resulted to be wrong when converting between language ie arab and latn.

* Lint error Fix

Lint error Fix